fix(MoneyVerifier): checkRange, checkDivisible 메서드 검증 타입 long 으로 수정

+++ src/main/java/lotto/verifier/MoneyVerifier.java
@@ -30,13 +30,13 @@ public class MoneyVerifier implements Verifier {
         }
     }
     private void checkRange(String input) {
-        int money = Integer.parseInt(input);
+        long money = Long.parseLong(input);
         if (money <= 0)
             throw new IllegalArgumentException(ExceptionMessage.IS_NOT_POSITIVE);
     }
 
     private void checkDivisible(String input) {
-        int money = Integer.parseInt(input);
+        long money = Long.parseLong(input);
         if (money % Constant.MONEY_UNIT != 0)
             throw new IllegalArgumentException(ExceptionMessage.IS_NOT_DIVISIBLE);
     }
